profile views,models and urls refactored

+++ user_profile/models.py
@@ -1,4 +1,4 @@
-from django.contrib.auth.models import AbstractUser, User
+from django.contrib.auth.models import User
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
@@ -8,9 +8,9 @@ class Profile(models.Model):
     bio = models.TextField(_('bio'), max_length=500, blank=True)
     rating = models.IntegerField(_("ratings"), default=500)
     games_played_count = models.FloatField(_("games_played_count"), default=0)
-    wins_count = models.FloatField(_("wins_count"), default=0)
-    losses_count = models.FloatField(_("losses_count"), default=0)
-    draw_count = models.FloatField(_("draw_count"), default=0)
+    wins_count = models.IntegerField(_("wins_count"), default=0)
+    losses_count = models.IntegerField(_("losses_count"), default=0)
+    draw_count = models.IntegerField(_("draw_count"), default=0)
     winning_percentage = models.FloatField(_("winning_percentage"), default=0)
     photo = models.ImageField(upload_to='media/images/', null=True, blank=True)
